And some more corrections
