+++ src/components/filter.js
@@ -1,7 +1,7 @@
-import {getMarkup} from "../utils";
+import {getMarkup} from "../utils/dom";
 import {filterElements} from "../data";
 
-const filterBlock = ({name, count = 0, isChecked = false} = {}) => {
+const filterBlock = ({name, count = 0, isChecked = false}) => {
   const id = name.toLowerCase();
   return `
   <input
